pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["travel_crew/src"]

[tool.hatch.build.targets.wheel]
packages = ["travel_crew"]

[project]
name = "travel-budget-estimator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.9.3",
    "crewai-tools>=1.9.3",
    "fastapi>=0.129.0",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[project.scripts]
travel-budget = "travel_crew.main:main"

[dependency-groups]
dev = [
    "hatchling>=1.28.0",
]
